release-1.6.0

+++ example/1004-move_arc_line.cc
@@ -38,7 +38,7 @@ int main(int argc, char **argv) {
         {300, 0, 200, 180, 0, 0},
     };
     fp32 radius = 0;
-    arm->set_pause_time(0.2);
+    arm->set_pause_time((fp32)0.2);
     for (int j = 0; j < 10; j++) {
         for (int i = 0; i < 6; i++) {
             ret = arm->set_position(poses[i], radius, false);
